+++ apps/api/src/services/impersonation/impersonation.module.ts
@@ -1,10 +1,12 @@
+import { SubscriptionModule } from '@ghostfolio/api/app/subscription/subscription.module';
+import { ConfigurationModule } from '@ghostfolio/api/services/configuration/configuration.module';
 import { ImpersonationService } from '@ghostfolio/api/services/impersonation/impersonation.service';
 import { PrismaModule } from '@ghostfolio/api/services/prisma/prisma.module';
 
 import { Module } from '@nestjs/common';
 
 @Module({
-  imports: [PrismaModule],
+  imports: [ConfigurationModule, PrismaModule, SubscriptionModule],
   providers: [ImpersonationService],
   exports: [ImpersonationService]
 })