+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added the write scopes to the access
+
 ### Changed
 
 - Improved the performance of the portfolio snapshot calculation by indexing the activities